+++ packages/model/src/standard/elements/RefrigeratorAndTemperatureControlledCabinetMode.ts
@@ -10,8 +10,8 @@ import { MatterDefinition } from "../MatterDefinition.js";
 import {
     ClusterElement as Cluster,
     AttributeElement as Attribute,
-    DatatypeElement as Datatype,
-    FieldElement as Field
+    FieldElement as Field,
+    DatatypeElement as Datatype
 } from "../../elements/index.js";
 
 export const RefrigeratorAndTemperatureControlledCabinetMode = Cluster({
@@ -23,10 +23,24 @@ export const RefrigeratorAndTemperatureControlledCabinetMode = Cluster({
 
     children: [
         Attribute({ name: "ClusterRevision", id: 0xfffd, type: "ClusterRevision", default: 2 }),
+
         Attribute({
-            name: "SupportedModes", id: 0x0, conformance: "M",
-            xref: { document: "cluster", section: "8.7.5" }
+            name: "SupportedModes", id: 0x0, type: "list", access: "R V", conformance: "M",
+            constraint: "2 to 255", quality: "F",
+
+            details: "This attribute shall contain the list of supported modes that may be selected for the CurrentMode " +
+                "attribute. Each item in this list represents a unique mode as indicated by the Mode field of the " +
+                "ModeOptionStruct." +
+                "\n" +
+                "Each entry in this list shall have a unique value for the Mode field. Each entry in this list shall " +
+                "have a unique value for the Label field.",
+
+            xref: { document: "cluster", section: "1.10.6.2" },
+            children: [
+                Field({ name: "entry", type: "RefrigeratorAndTemperatureControlledCabinetMode.ModeOptionStruct" })
+            ]
         }),
+
         Attribute({ name: "CurrentMode", id: 0x1, conformance: "M", xref: { document: "cluster", section: "8.7.5" } }),
         Attribute({ name: "StartUpMode", id: 0x2, conformance: "P", xref: { document: "cluster", section: "8.7.5" } }),
         Attribute({ name: "OnMode", id: 0x3, conformance: "P", xref: { document: "cluster", section: "8.7.5" } }),
@@ -38,7 +52,14 @@ export const RefrigeratorAndTemperatureControlledCabinetMode = Cluster({
                 "\n" +
                 "At least one entry in the SupportedModes attribute shall include the Auto mode tag in the ModeTags " +
                 "field list.",
-            xref: { document: "cluster", section: "8.7.4.1" }
+            xref: { document: "cluster", section: "8.7.4.1" },
+
+            children: [Field({
+                name: "ModeTags", type: "list",
+                children: [
+                    Field({ name: "entry", type: "RefrigeratorAndTemperatureControlledCabinetMode.ModeTagStruct" })
+                ]
+            })]
         }),
 
         Datatype({
@@ -56,6 +77,11 @@ export const RefrigeratorAndTemperatureControlledCabinetMode = Cluster({
                     xref: { document: "cluster", section: "8.7.6.1.2" }
                 })
             ]
+        }),
+
+        Datatype({
+            name: "ModeTagStruct", type: "ModeTagStruct",
+            children: [Field({ name: "Value", type: "RefrigeratorAndTemperatureControlledCabinetMode.ModeTag" })]
         })
     ]
 })