+++ CHANGELOG.md
@@ -9,6 +9,11 @@ The main work (all changes without a GitHub username in brackets in the below li
 	## __WORK IN PROGRESS__
 -->
 
+## __WORK IN PROGRESS__
+
+-   @matter/model
+    - Enhancement: The generated model now adds extensions cluster elements as required to enable Matter enums that extend parent enums
+
 ## 0.11.8 (2024-11-29)
 
 -   @matter/protocol